+++ x/oracle/utils/assets.go
@@ -1,14 +1,14 @@
 package utils
 
 const (
-	MicroUsdcDenom = "uusdc"
-	MicroKiiDenom  = "akii"
-	MicroAtomDenom = "uatom"
-	MicroEthDenom  = "ueth"
-	MicroBtcDenom  = "ubtc"
-	MicroSolDenom  = "usol"
-	MicroXrpDenom  = "uxrp"
-	MicroBnbDenom  = "ubnb"
-	MicroUsdtDenom = "uusdt"
-	MicroTrxDenom  = "utrx"
+	UsdcDenom = "usdc"
+	KiiDenom  = "kii"
+	AtomDenom = "atom"
+	EthDenom  = "eth"
+	BtcDenom  = "btc"
+	SolDenom  = "sol"
+	XrpDenom  = "xrp"
+	BnbDenom  = "bnb"
+	UsdtDenom = "usdt"
+	TrxDenom  = "trx"
 )